infrastructure/account_b/lambda/s3_redshift_lambda/requirements.txt
psycopg2==2.7.7
python-dotenv==0.19.0